requirements.txt
Flask==3.0.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-Admin
Flask-Login
folium==0.20.0
gunicorn>=21.2
python-dotenv
psycopg2-binary
